chore(tests): extract helper methods to MockFactory
- Extract CreateSummary method for benchmarks with metrics
- Extract FakeMetricDescriptor into separate file
